Integration of license-maven-plugin-git for Git-related Functionality
Motivation:

The license-gradle-plugin was not maintained, so we decided to take ownership and enhance it ourselves. The original repository URL is [license-gradle-plugin](https://github.com/hierynomus/license-gradle-plugin).

Modification:

Add license-maven-plugin-git as a dependency and configure it for git operations. The integrated module of license-maven-plugin-git can be found at [license-maven-plugin-git](https://github.com/mathieucarbou/license-maven-plugin/tree/master/license-maven-plugin-git).

Result:

Enhanced license-gradle-plugin with git support for managing licenses in the project.